+++ packages/common-ethereum/src/project/types.ts
@@ -25,15 +25,11 @@ export {
 export type IEthereumProjectManifest = IProjectManifest<SubqlDatasource>;
 
 export enum SubqlEthereumHandlerKind {
-  FlareBlock = 'flare/BlockHandler',
-  FlareCall = 'flare/TransactionHandler',
-  FlareEvent = 'flare/LogHandler',
   EthBlock = 'ethereum/BlockHandler',
   EthCall = 'ethereum/TransactionHandler',
   EthEvent = 'ethereum/LogHandler',
 }
 
 export enum SubqlEthereumDatasourceKind {
-  FlareRuntime = 'flare/Runtime',
   EthRuntime = 'ethereum/Runtime',
 }