+++ packages/node/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.6.4] - 2024-06-12
+### Changed
+- Update `@subql/utils`
+
+## [4.6.3] - 2024-06-12
+### Changed
+- Update polkadot dependencies to 11.2.1 (#2440)
+
 ## [4.6.2] - 2024-06-06
 ### Fixed
 - Fix import monitor service from node-core
@@ -1216,7 +1224,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/4.6.2...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node/4.6.4...HEAD
+[4.6.4]: https://github.com/subquery/subql/compare/node/4.6.3...node/4.6.4
+[4.6.3]: https://github.com/subquery/subql/compare/node/4.6.2...node/4.6.3
 [4.6.2]: https://github.com/subquery/subql/compare/node/4.6.1...node/4.6.2
 [4.6.1]: https://github.com/subquery/subql/compare/node/4.6.0...node/4.6.1
 [4.6.0]: https://github.com/subquery/subql/compare/node/4.5.1...node/4.6.0